+++ react-components/src/components/CookiesModal/CookiesModal.jsx
@@ -12,14 +12,18 @@ export function CookiesModal(props) {
   const focusTrap = React.useRef(false);
 
   function handleAcceptAllCookies(event) {
-    CookieManager.acceptAllCookiesAndShowSuccess(event);
-    // window.location.reload(); // disabled to avoid the page reloading before the user can see the banner
-    setIsOpen(false);
+    CookieManager.acceptAllCookiesAndShowSuccess(event)
+
+    window.dataLayer = window.dataLayer || []
+    window.dataLayer.push({ event: 'cookies_policy_accept' })
+    window.dataLayer.push({ event: 'gtm.dom' })
+
+    setIsOpen(false)
   }
 
   const handleFocusChange = (evt) => {
     if (focusTrap.current) {
-      let modal = evt.target.closest('.ReactModal__Content');
+      const modal = evt.target.closest('.ReactModal__Content');
       if ((!modal || (modal == evt.target)) && firstLink) {
         firstLink.focus();
       }
@@ -49,8 +53,9 @@ export function CookiesModal(props) {
         >cookies to collect information</a> about how you use great.gov.uk. We use this information to make the website work as well as possible and improve government services.
       </p>
       <div className={styles.buttonContainer}>
-        <a className={`${styles.button} button`} href="#" onClick={handleAcceptAllCookies}>Accept all cookies</a>
-        <span className={styles.buttonSeperator}></span>
+        <a className={`${styles.button} button`} href="#"
+onClick={handleAcceptAllCookies}>Accept all cookies</a>
+        <span className={styles.buttonSeperator} />
         <a className={`${styles.button} button`} href={props.preferencesUrl + window.location.search}>Set cookie preferences</a>
       </div>
     </Modal>